feat: start implementing secret instance config methods
